requirements.txt
actionlib==1.12.1
angles==1.9.12
appdirs==1.4.4
apturl==0.5.2
asgiref==3.3.4
asn1crypto==0.24.0
beautifulsoup4==4.6.0
blinker==1.4
bondpy==1.8.5
Brlapi==0.6.6
camera-calibration==1.15.0
camera-calibration-parsers==1.11.13
catkin==0.7.29
certifi==2018.1.18
chardet==3.0.4
click==7.1.2
command-not-found==0.3
controller-manager==0.18.3
controller-manager-msgs==0.18.3
controller-manager-tests==0.18.3
cryptography==2.1.4
cupshelpers==1.0
cv-bridge==1.13.0
defer==1.0.6
diagnostic-analysis==1.9.7
diagnostic-updater==1.9.7
distlib==0.3.1
distro-info===0.18ubuntu0.18.04.1
Django==3.2.1
django-filter==2.4.0
djangorestframework==3.12.4
dynamic-reconfigure==1.6.3
dynamixel-sdk==3.7.31
feedparser==5.2.1
filelock==3.0.12
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.14
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
image-geometry==1.13.0
importlib-metadata==4.0.1
importlib-resources==5.1.2
interactive-markers==1.11.5
joblib==1.0.1
joint-state-publisher==1.12.15
joint-state-publisher-gui==1.12.15
kdl-parser-py==1.13.1
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
laser-geometry==1.6.7
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.4
MarkupSafe==1.0
message-filters==1.14.10
moveit-ros-planning-interface==1.0.7
mpmath==1.1.0
netifaces==0.10.4
nltk==3.5
numpy==1.19.4
oauth==1.0.1
oauthlib==2.0.6
olefile==0.45.1
pexpect==4.2.1
Pillow==8.2.0
pr2-controller-manager==1.8.18
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.6.1
python-debian==0.1.32
python-qt-binding==0.4.3
pytz==2021.1
pyxdg==0.25
PyYAML==3.12
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
regex==2020.11.13
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
resource-retriever==1.12.6
rosbag==1.14.10
rosclean==1.14.9
rosgraph==1.14.10
roslaunch==1.14.10
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.10
rosmake==1.14.9
rosmaster==1.14.10
rosmsg==1.14.10
rosnode==1.14.10
rosparam==1.14.10
rospy==1.14.10
rosservice==1.14.10
rostest==1.14.10
rostopic==1.14.10
rosunit==1.14.9
roswtf==1.14.10
rqt-action==0.4.9
rqt-console==0.4.9
rqt-controller-manager==0.18.3
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-joint-trajectory-controller==0.17.2
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-py-common==0.5.2
rqt-py-console==0.4.8
rqt-runtime-monitor==0.5.7
rqt-rviz==0.6.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-topic==0.4.11
rviz==1.13.16
SecretStorage==2.3.1
seed-smartactuator-sdk==0.0.5
sensor-msgs==1.12.8
simplejson==3.13.2
six==1.15.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.5
speedtest-cli==2.0.0
sqlparse==0.4.1
srdfdom==0.5.2
sympy==1.7.1
system-service==0.3
systemd-python==234
tf==1.12.1
tf-conversions==1.12.1
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
topic-tools==1.14.10
tqdm==4.60.0
typing-extensions==3.10.0.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urdfdom-py==0.4.5
urllib3==1.22
usb-creator==0.3.3
virtualenv==20.4.4
wadllib==1.3.2
webencodings==0.5
xacro==1.13.9
xkit==0.0.0
zipp==3.4.1
zope.interface==4.3.2
